+++ R/mod_mc_mc.R
@@ -188,6 +188,7 @@ mod_mc_mc_server <- function(id, dataset, metacell_types, cell_type_colors, gene
             # Projection plots
             output$plot_mc_proj_2d <- render_2d_plotly(input, output, session, dataset, metacell_types, cell_type_colors, gene_modules, globals, groupA = groupA, groupB = groupB, source = "proj_mc_plot")
 
+            diff_expr_auto_update_globals(mc_mc_gene_scatter_df, globals)
 
             # metacell click observers
             metacell_click_observer("proj_manifold_plot", session)