+++ NEWS.md
@@ -1,3 +1,7 @@
+# MCView 0.2.44
+
+* Added "User DE genes" button to the 'Markers' tab, that uses the significantly differentially expressed genes from the 'Diff. Expr.' tab.
+
 # MCView 0.2.43
 
 * Added "Highlight selected genes" option to the 'Markers' tab.